Transform CARLA into true reinforcement learning agent
- Change grind_size from float to int for realistic discrete grinder settings (1-30)
- Remove all hardcoded coffee expertise and bean aging logic
- Replace smart baselines with neutral espresso defaults (grind=15, volume=40ml, dose=18g)
- Agent now learns optimal parameters through experience rather than following predetermined rules
- Update environment bounds and discretization for integer grind sizes
- Fix CLI and export displays to show integer grind values
- Enable true discovery of brewing relationships through user feedback

This transforms CARLA from a biased expert system into a pure RL agent that can discover
personalized and potentially surprising optimal brewing parameters.

+++ carla/models.py
@@ -13,7 +13,7 @@ MAX_RATING = 10
 @dataclass
 class BrewAction:
     """Action parameters for brewing espresso."""
-    grind_size: float  # Most important parameter
+    grind_size: int  # Most important parameter (1-30, discrete grinder settings)
     brew_volume: float  # ml
     coffee_dose: float  # g
 
@@ -21,7 +21,7 @@ class BrewAction:
 @dataclass
 class BrewState:
     """State information for brewing context."""
-    is_first_brew: bool
+    is_first_brew: bool  # First brew after machine startup (warming up)
     days_since_roast: int
 
     @classmethod
